refactor: input 기능별로 분리
