+++ src/components/form/form-label.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+type FormLabelProps = {
+  htmlFor: string;
+  children: React.ReactNode;
+};
+
+const FormLabel = ({ htmlFor, children }: FormLabelProps) => {
+  return (
+    <label htmlFor={htmlFor} className='block text-sm font-medium text-gray-700'>
+      {children}
+    </label>
+  );
+};
+
+export default FormLabel;
